pyproject.toml
[project]
name = "meetup-graphql"
version = "0.1.0"
description = "A project for the Django London Meetup group using Python."
authors = [{ name = "Ester Beltrami", email = "beltrami.ester@gmail.com" }]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.2",
    "openpyxl>=3.1.2",
    "pyjwt>=2.9.0",
    "cryptography>=43.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }
keywords = ["django", "diversity", "meetup", "analysis", "openai", "flask"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License"
]
homepage = "https://github.com/estyxx/meetup-diversity"
repository = "https://github.com/estyxx/meetup-diversity"
documentation = "https://github.com/estyxx/meetup-diversity#readme"

[project.optional-dependencies]
linters = ["ruff>=0.2.1", "pre-commit>=3.6.1", "mypy>=1.8.0"]

[tool.pdm.scripts]
_.env_file = ".env"
export = { call = "main:main" }

[tool.pdm]
distribution = false
plugins = ["sync-pre-commit-lock"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort

]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_return_any = true
